Changed ApiSchemaBuilder file name to Index
